+++ src/interface/src/app/maplibre-map/draw.service.ts
@@ -16,9 +16,9 @@ import {
 } from 'rxjs';
 import { GeoJSON } from 'geojson';
 import booleanWithin from '@turf/boolean-within';
-import flatten from '@turf/flatten';
 import { HttpClient } from '@angular/common/http';
 import { FeatureService } from '../features/feature.service';
+import { flattenMultipolygons } from '../plan/plan-helpers';
 
 export type DrawMode = 'polygon' | 'select' | 'none';
 
@@ -257,22 +257,6 @@ export class DrawService {
     return this._totalAcres$.value;
   }
 
-  flattenPolygons(featuresArray : any[]) {
-    const justPolygons: any[] = [];
-    featuresArray.forEach((f: any) => {
-      if (f.geometry.type === 'Polygon') {
-        justPolygons.push(f);
-      }else if (f.geometry.type === 'MultiPolygon') {
-        const flattened = flatten(f);
-        flattened.features.forEach((flat) => {
-          if ((flat.geometry as any).type === 'Polygon') {
-            justPolygons.push(flat);
-          }
-        });
-      }
-    })
-    return justPolygons;
-  }
 
   addGeoJSONFeature(shape: any) {
     const featuresArray = shape.features.map((feature: any) => ({
@@ -286,7 +270,7 @@ export class DrawService {
         mode: 'polygon',
       },
     }));
-    const flatFeatures = this.flattenPolygons(featuresArray);
+    const flatFeatures = flattenMultipolygons(featuresArray);
     this._terraDraw?.addFeatures(flatFeatures);
     this.updateTotalAcreage();
     this._terraDraw?.setMode('select'); // should be in select mode to add
